+++ paramdb/_param_data/_dataclasses.py
@@ -34,8 +34,7 @@ class _ParamDataclass(ParamData):
 
     @property
     @abstractmethod
-    def last_updated(self) -> datetime | None:  # pragma: no cover
-        ...
+    def last_updated(self) -> datetime | None: ...
 
     def to_dict(self) -> dict[str, Any]:
         if is_dataclass(self):
@@ -49,6 +48,8 @@ class _ParamDataclass(ParamData):
 
 class Param(_ParamDataclass):
     """
+    Subclass of :py:class:`ParamData`.
+
     Base class for parameters. Subclasses are automatically converted to dataclasses.
     For example::
 
@@ -85,6 +86,8 @@ class Param(_ParamDataclass):
 
 class Struct(_ParamDataclass):
     """
+    Subclass of :py:class:`ParamData`.
+
     Base class for parameter structures. Subclasses are automatically converted to
     dataclasses. For example::