+++ tests/helpers.py
@@ -6,9 +6,11 @@
 
 from __future__ import annotations
 from typing import Any
-from dataclasses import field
+from collections.abc import Iterator
+from dataclasses import dataclass, field
+from contextlib import contextmanager
 import time
-from astropy.units import Quantity  # type: ignore
+from astropy.units import Quantity  # type: ignore # pylint: disable=import-error
 from paramdb import ParamData, Param, Struct, ParamList, ParamDict
 
 DEFAULT_NUMBER = 1.23
@@ -47,14 +49,29 @@ class CustomParamDict(ParamDict[Any]):
     """Custom parameter dictionary subclass."""
 
 
-def sleep_for_datetime() -> None:
-    """
-    Wait for a short amount of time so that following calls to `datetime.now()` are
-    different than those that come before.
+@dataclass
+class Times:
+    """Start and end times captured by ``capture_start_end_times()``."""
 
-    This is necessary because a `datetime` object is only precise to microseconds (see
-    https://docs.python.org/3/library/datetime.html#datetime-objects), whereas modern
-    computers execute instructions faster than this. So without waiting, it is difficult
-    to ensure that something is using `datetime.now()`.
+    start: float
+    end: float = 0
+
+
+@contextmanager
+def capture_start_end_times(wait_time: float = 10e-6) -> Iterator[Times]:
     """
-    time.sleep(100e-6)  # Wait for 100 microseconds
+    Record the start and end times within a with statement using this context manager.
+
+    Also waits for a short amount of time after setting the start time and before
+    setting the end time so that following calls to ``datetime.now()`` are different
+    that then one being targeted. This is necessary because a ``datetime`` object is
+    only precise to microseconds (see
+    https://docs.python.org/3/library/datetime.html#datetime-objects), whereas computers
+    execute instructions faster. So without waiting a few microseconds, it is difficult
+    to verify that something is using the current time.
+    """
+    times = Times(time.time())
+    time.sleep(wait_time)
+    yield times
+    time.sleep(wait_time)
+    times.end = time.time()